+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/ingestion/database/DirtyEventSQLiteDao.kt
@@ -13,7 +13,8 @@ import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 
 class DirtyEventSQLiteDao(
     private val databaseManager: DatabaseManager,
-    private val logger: Logger
+    private val logger: Logger,
+    private val eventTypeConverter: EventTypeConverters
 ) : DirtyEventDao, DatabaseTable {
     override val tableName = "DirtyEvents"
     override val columns: Map<String, String>
@@ -38,7 +39,7 @@ class DirtyEventSQLiteDao(
         return retrievedDataList.map { retrievedData ->
             DirtyMeetingEventItem(
                 retrievedData[idColumnName] as String,
-                EventTypeConverters.toMeetingEvent(retrievedData[dataColumnName] as String),
+                eventTypeConverter.toMeetingEvent(retrievedData[dataColumnName] as String),
                 retrievedData[ttlColumnName] as Long
             )
         }
@@ -52,7 +53,7 @@ class DirtyEventSQLiteDao(
         return databaseManager.insert(tableName, dirtyEvents.map { dirtyEvent ->
             ContentValues().apply {
                 put(idColumnName, dirtyEvent.id)
-                put(dataColumnName, EventTypeConverters.fromMeetingEvent(dirtyEvent.data))
+                put(dataColumnName, eventTypeConverter.fromMeetingEvent(dirtyEvent.data))
                 put(ttlColumnName, dirtyEvent.ttl)
             }
         })